scorer: 将编码器 dtype 从 bfloat16 改为 float32
修改 PickScore/CLIP/HPSv2 构造中的 dtype 参数为 torch.float32，以避免在某些设备上 bf16 支持不完整导致的不兼容问题。

+++ reward_models/camera_normal_scorer/scorer.py
@@ -39,7 +39,7 @@ class CameraNormalScorer:
                 device=device,
                 model_id=self.cfg.pickscore_model_id,
                 processor_id=self.cfg.pickscore_processor_id,
-                dtype=torch.bfloat16,
+                dtype=torch.float32,
                 similarity_type="cls",
             )
         elif enc_name == "clip":
@@ -48,7 +48,7 @@ class CameraNormalScorer:
                 device=device,
                 model_id=self.cfg.clip_model_id,
                 processor_id=self.cfg.clip_processor_id,
-                dtype=torch.bfloat16,
+                dtype=torch.float32,
                 similarity_type="cls",
             )
         elif enc_name == "hpsv2":
@@ -56,7 +56,7 @@ class CameraNormalScorer:
             self.encoder = HPSV2ImageEncoder(  # 形状: 编码器
                 device=device,
                 ckpt_path=self.cfg.hpsv2_ckpt_path,
-                dtype=torch.bfloat16,
+                dtype=torch.float32,
                 similarity_type="cls",
             )
         elif enc_name == "dino_v2":
